use size_t and notebook added

+++ shuffle.c
@@ -1,9 +1,8 @@
-#include <stdlib.h> //rand
 #include "shuffle.h"
 
-void shuffle(double *x, int n)
+void shuffle(double *x, size_t n)
 {
-    int j,k;
+    size_t j,k;
     double tmp;
     for(j=0; j<n; j++){
         k = (int)rand() % n ;
