+++ test.c
@@ -4,12 +4,12 @@
 
 int main(){
     double x[] = {1,2,3,4,5,6,7,8,9,0};
-    int n = 10;
+    size_t n = 10;
     
     srand(42); //set seed
     shuffle(x,n); //permute
     
-    for (int i=0; i<n; i++){
+    for (size_t i=0; i<n; i++){
         printf("%.f ", x[i]);
     }
     printf("\n");
